fix: Memisahkan komponen
komponen yang memakan banyak baris kode dipisahkan menjadi per file untuk kemudahan maintenance
